+++ src/ezdxf/entities/shape.py
@@ -116,4 +116,5 @@ class Shape(DXFGraphic):
             dxf.thickness = ocs.transform_thickness(dxf.thickness)
 
         dxf.extrusion = ocs.new_extrusion
+        self.post_transform(m)
         return self